+++ src/stores/entities/menu.entity.ts
