+++ src/stores/entities/store.entity.ts
@@ -2,7 +2,7 @@ import { Column, CreateDateColumn, Entity, JoinColumn, ManyToOne, OneToMany, One
 import { StoreCategory } from "./store-category.enum"
 import { Review } from "src/reviews/entites/review.entity"
 import { StoreRequest } from "../../store-requests/entities/store-request.entity"
-import { Menu } from "src/menus/entities/menu.entity"
+import { Menu } from "src/stores/entities/menu.entity"
 import { ManagerRequest } from "src/manager-requests/entities/manager-requests.entity"
 import { User } from "src/users/entities/user.entity"
 import { Event } from "./event.entity"
