+++ src/stores/stores.controller.ts
@@ -18,6 +18,11 @@ import { ReadEventDTO } from './DTO/read-event.dto'
 import { ReadAllEventsDTO } from './DTO/read-all-events.dto'
 import { UpdateEventDTO } from './DTO/update-event.dto'
 import { AuthenticatedRequest } from 'src/auth/interfaces/authenticated-request.interface'
+import { CreateMenuDTO } from './DTO/create-menu.dto'
+import { Menu } from './entities/menu.entity'
+import { ReadAllMenusDTO } from './DTO/read-all-menus.dto'
+import { ReadMenuDTO } from './DTO/read-menu.dto'
+import { UpdateMenuDTO } from './DTO/update-menu.dto'
 
 @Controller('api/stores')
 @UseGuards(AuthGuard('jwt'), RolesGuard) // JWT인증, roles guard 적용
@@ -258,5 +263,63 @@ export class StoresController {
         const readReviewDTOs = foundReviews.map(review => new ReadReviewDTO(review))
 
         return new ApiResponseDTO(true, HttpStatus.OK, "Store Reviews Retrieved Successfully", readReviewDTOs)
+    }
+
+
+    // ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ메뉴 관련 기능 시작ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ
+
+    // CREATE - 새로운 메뉴 등록
+    // 미구현: logger
+    @Post('/')
+    @Roles(UserRole.MANAGER)
+    async createMenu(
+        @Param('store_id') store_id: number,
+        @Body() createMenuDTO: CreateMenuDTO): Promise<ApiResponseDTO<Menu>> {
+        await this.storesService.createMenu(store_id, createMenuDTO)
+
+        return new ApiResponseDTO(true, HttpStatus.CREATED, 'Menu Registered Successfully!')
+    }
+
+    // READ[1] - 해당 가게 모든 메뉴 조회
+    // 미구현: logger
+    @Get('/')
+    async readMenusByStore(@Param('store_id') store_id: number): Promise<ApiResponseDTO<ReadAllMenusDTO[]>> {
+        const menus: Menu[] = await this.storesService.readMenusByStore(store_id)
+        const readAllMenusDTO = menus.map(menu => new ReadAllMenusDTO(menu))
+
+        return new ApiResponseDTO(true, HttpStatus.OK, 'Successfully Retrieved All Menus!', readAllMenusDTO)
+    }
+
+    // READ[2] - 해당 가게 특정 메뉴 조회
+    // 미구현: logger
+    @Get('/:menu_id')
+    async readMenuById(
+        @Param('menu_id') menu_id: number): Promise<ApiResponseDTO<ReadMenuDTO>> {
+        const foundMenu: Menu = await this.storesService.readMenuById(menu_id)
+
+        return new ApiResponseDTO(true, HttpStatus.OK, 'Successfully Retrieved Event!', new ReadMenuDTO(foundMenu))
+    }
+
+    // UPDATE - by menu_id
+    // 미구현: logger
+    @Put('/:menu_id')
+    @Roles(UserRole.MANAGER)
+    async updateMenuById(
+        @Param('menu_id') menu_id: number,
+        @Body() updateMenuDTO: UpdateMenuDTO): Promise<ApiResponseDTO<void>> {
+        await this.storesService.updateMenuById(menu_id, updateMenuDTO)
+
+        return new ApiResponseDTO(true, HttpStatus.NO_CONTENT, 'Menu Updated Successfully!')
+    }
+
+    // DELETE - by menu_id
+    // 미구현: logger
+    @Delete('/:menu_id')
+    @Roles(UserRole.MANAGER)
+    async deleteMenuById(
+        @Param('menu_id') menu_id: number): Promise<ApiResponseDTO<void>> {
+        await this.storesService.deleteMenuById(menu_id)
+
+        return new ApiResponseDTO(true, HttpStatus.NO_CONTENT, 'Menu Deleted Successfully!')
     }
 }
